+++ tests/utils/test_train_utils.py
@@ -1,8 +1,16 @@
+from typing import Any, Dict
+
 import numpy as np
+import pytest
 
 import rasa.utils.train_utils as train_utils
+from rasa.core.policies.ted_policy import TEDPolicy
 from rasa.nlu.constants import NUMBER_OF_SUB_TOKENS
 from rasa.nlu.tokenizers.tokenizer import Token
+from rasa.shared.nlu.constants import (
+    SPLIT_ENTITIES_BY_COMMA_DEFAULT_VALUE,
+    SPLIT_ENTITIES_BY_COMMA,
+)
 
 
 def test_align_token_features():
@@ -26,3 +34,31 @@ def test_align_token_features():
     assert np.all(actual_features[0][3] == np.mean(token_features[0][3:5], axis=0))
     # embedding is split into 4 sub-tokens
     assert np.all(actual_features[0][4] == np.mean(token_features[0][5:10], axis=0))
+
+
+@pytest.mark.parametrize(
+    "split_entities_config, expected_initialized_config",
+    [
+        (
+            SPLIT_ENTITIES_BY_COMMA_DEFAULT_VALUE,
+            {SPLIT_ENTITIES_BY_COMMA: SPLIT_ENTITIES_BY_COMMA_DEFAULT_VALUE},
+        ),
+        (
+            {"address": False, "ingredients": True},
+            {
+                "address": False,
+                "ingredients": True,
+                SPLIT_ENTITIES_BY_COMMA: SPLIT_ENTITIES_BY_COMMA_DEFAULT_VALUE,
+            },
+        ),
+    ],
+)
+def test_init_split_entities_config(
+    split_entities_config: Any, expected_initialized_config: Dict[(str, bool)],
+):
+    assert (
+        train_utils.init_split_entities(
+            split_entities_config, SPLIT_ENTITIES_BY_COMMA_DEFAULT_VALUE
+        )
+        == expected_initialized_config
+    )
